manifest: Drop gnome-builder
It requires clang, and we only currently kind of hackily "maintain"
LLVM in order to use Mesa/llvmpipe.

If someone steps up to maintain it, we can readd Builder.

As is now, it's been tagged for a while, but the tagged version
doesn't actually build because glib changed to add the schema
validation.