patches: Use a big hammer on the webkitgtk4 patch
At this point I frankly have no clue what is going on. Upgrading WebKit
definitely causes this error, but it can't be to any change in the CMake
build system, as the error is that CMakeLists.txt cannot be found....

The srcdir is one directory higher than the builddir, but for some
reason it's not working, so try forcing it one directory higher. This
command ought to be telling cmake to look one directory too high, but
hey, maybe it will work....