patches: Fix webkitgtk-cmake-buildapi.patch
We added -DENABLE_FTL_JIT=OFF to this patch without updating the line
numbers, so the last line in the patch file got silently truncated, and
the last line was where we pass the srcdir to cmake.

At least I think that's what happened. We'll find out if this builds!