Untag mesa and get it building against LLVM 3.3
In order to find the 3.3 of LLVM that is build in our yocto base,
it's necessary to patch mesa's configure.ac so that it can be
made to look in /usr/bin/llvm3.3 for llvm-config. It's also necessary
to pass a --host value, since llvm config is installed as
x86_64-gnomeostree-linux. If we start building on other architectures
this will need to adjustment.

(Before the Yocto rebase, we had a LLVM 3.4 custom recipe, that stopped
working; LLVM 3.3 and 3.4 are both sufficient to get llvmpipe working, but
not new enough for some other parts of Mesa like amdgpu, the radv Vulkan support,
etc. If we want those, we'll need to update to something considerably
newer.)