Attempt at building LLVM inside Continuous
Various projects are starting to depend on versions of LLVM newer than
the 3.3 version that we have in our Yocto base.

Yocto is stuck to LLVM 3.3 because after that release LLVM switched to
CMake and dropped autotools. Finding a modern recipe for building LLVM
in our Yocto base is a lost cause, since everyone seems to be building
either Clang or embedded copies of LLVM, these days.