bot: Notify failures for components with an irc_room
The existing code looks at the built modules, and if the previous status
was a failure, this will trigger a notification on the specified IRC
channel.

In order to be relatively more useful as a form of notification, we
should look at which module failed to build, and if it specifies an IRC
channel, we should send a notification there as well.