Do yet another attempt at a UTF-8 locale
We build inside a Yocto build root, and since Yocto does not ship with a
patched GNU libc to provide a C.UTF-8 locale (like Fedora and Debian do)
we need to use a valid locale otherwise we're going to break everything.

The image we build defaults to en_US.UTF-8, so let's use the same locale
during the build.

At some point, with a newer version of Yocto and a newer GNU libc that
solves this:

  https://sourceware.org/glibc/wiki/Proposals/C.UTF-8

we're going to switch to C.UTF-8.