Provide fallback values for CC and CXX
If CC and CXX are unset, distutils will fail miserably and use the C
compiler that was used to build it — which, in the case of Continuous,
is the C compiler used inside the Yocto base, and which tries to access
the system includes outside of the sandbox.

Let's fall back to 'gcc' and 'g++' until the glorious day when we can
take distutils and dump it in a very dark and deep pit.